LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-in-body-onerror-expected.txt

CONSOLE MESSAGE: line 19: SyntaxError: Unexpected token ')'
CONSOLE MESSAGE: line 16: SyntaxError: Unexpected end of script

PASS window.onerror - compile error in <body onerror> 
